# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "An API used to manage payments with friends for home poker games :)",
        "title": "Poker Finance API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/games": {
            "post": {
                "description": "Create a new game",
                "tags": [
                    "games"
                ],
                "summary": "Create a new game",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.Game"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/games/{id}": {
            "get": {
                "description": "Get a game by ID",
                "tags": [
                    "games"
                ],
                "summary": "Get a game by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.Game"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a game by ID and all related played games",
                "tags": [
                    "games"
                ],
                "summary": "Delete a game by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payment-details": {
            "post": {
                "description": "Create new payment details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment-details"
                ],
                "summary": "Create new payment details",
                "parameters": [
                    {
                        "description": "Payment Details Input",
                        "name": "paymentDetails",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.CreatePaymentDetailsInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.PaymentDetails"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payment-details/{id}": {
            "get": {
                "description": "Get payment details by ID",
                "tags": [
                    "payment-details"
                ],
                "summary": "Get payment details by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment Details ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.PaymentDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update payment details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment-details"
                ],
                "summary": "Update payment details by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment Details ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payment Details Input",
                        "name": "paymentDetails",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.UpdatePaymentDetailsInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.PaymentDetails"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions": {
            "post": {
                "description": "Create a new transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Create a new transaction",
                "parameters": [
                    {
                        "description": "Transaction Input",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.CreateTransactionInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.Transaction"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions/{id}": {
            "put": {
                "description": "Update transaction status by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Update transaction status by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status Update Input",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.UpdateTransactionStatusInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.Transaction"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "backend_internal_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "backend_internal_models.Game": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "backend_internal_models.PaymentDetails": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "payee": {
                    "$ref": "#/definitions/backend_internal_models.User"
                },
                "payeeID": {
                    "type": "string"
                },
                "payer": {
                    "$ref": "#/definitions/backend_internal_models.User"
                },
                "payerID": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "timeCompleted": {
                    "type": "string"
                },
                "timeSubmitted": {
                    "type": "string"
                }
            }
        },
...
        "backend_internal_models.Transaction": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "paymentDetails": {
                    "$ref": "#/definitions/backend_internal_models.PaymentDetails"
                },
                "paymentDetailsID": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/backend_internal_models.User"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "backend_internal_models.User": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "internal_api_handlers.CreatePaymentDetailsInput": {
            "type": "object",
            "required": [
                "amount",
                "payee_id",
                "payer_id",
                "status"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                },
                "payee_id": {
                    "type": "string"
                },
                "payer_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "internal_api_handlers.CreateTransactionInput": {
            "type": "object",
            "properties": {
                "payment_details_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_handlers.UpdatePaymentDetailsInput": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                },
                "payee_id": {
                    "type": "string"
                },
                "payer_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "time_completed": {
                    "type": "string"
                }
            }
        },
        "internal_api_handlers.UpdateTransactionStatusInput": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}